SWC-6520 - Extract new component QueryWrapperLoadingScreen and re-use in StandaloneQueryWrapper
