+++ packages/synapse-react-client/src/components/QueryWrapperPlotNav/FilterAndView.tsx
@@ -2,15 +2,9 @@ import React from 'react'
 import CardContainer from '../CardContainer'
 import SynapseTable, { SynapseTableProps } from '../SynapseTable'
 import { CardConfiguration } from '../CardContainerLogic'
-import { useQueryContext } from '../QueryContext'
 import LastUpdatedOn from './LastUpdatedOn'
-import { Box, LinearProgress, Typography } from '@mui/material'
 import { useDeepCompareMemoize } from 'use-deep-compare-effect'
-import { useAtomValue } from 'jotai'
-import {
-  isLoadingNewBundleAtom,
-  tableQueryDataAtom,
-} from '../QueryWrapper/QueryWrapper'
+import QueryWrapperLoadingScreen from '../QueryWrapper/QueryWrapperLoadingScreen'
 
 export type FilterAndViewProps = {
   tableConfiguration?: SynapseTableProps
@@ -28,30 +22,10 @@ const FilterAndView = (props: FilterAndViewProps) => {
         }
       : undefined,
   )
-  const queryContext = useQueryContext()
-  const { asyncJobStatus } = queryContext
-  const isLoadingNewBundle = useAtomValue(isLoadingNewBundleAtom)
-  const data = useAtomValue(tableQueryDataAtom)
+
   return (
     <div className={`FilterAndView`}>
-      {!data && isLoadingNewBundle && (
-        <Box sx={{ mt: 15, mx: 15 }}>
-          <LinearProgress />
-          {asyncJobStatus?.progressMessage && (
-            <Typography
-              variant={'smallText1'}
-              sx={{
-                color: 'grey.600',
-                fontSize: '10px',
-                textAlign: 'center',
-                my: 1,
-              }}
-            >
-              {asyncJobStatus?.progressMessage}
-            </Typography>
-          )}
-        </Box>
-      )}
+      <QueryWrapperLoadingScreen />
       {tableConfiguration ? <SynapseTable {...tableConfiguration} /> : <></>}
       {cardConfiguration ? <CardContainer {...cardConfiguration} /> : <></>}
       <LastUpdatedOn />
